+++ CHANGELOG.md
@@ -3,11 +3,16 @@ All notable changes to this project will be documented in this file.
 
 See [DocuSign Support Center](https://support.docusign.com/en/releasenotes/) for Product Release Notes.
 
-## [v2.11.0-rc1] - eSignature API v2-21.2.00.00 - 2021-07-13
+## [v2.12.0.rc1] - ESignature API v2-21.2.02.00 - 2021-08-30
+### Changed
+- Added support for version v2-21.2.02.00 of the DocuSign ESignature API.
+- Updated the SDK release version.
+
+## [v2.11.0] - eSignature API v2-21.1.02.00 - 2021-07-13
 ### Added
 - Added new models and properties.
 ### Changed
-- Added support for version v2-21.2.00.00 of the DocuSign eSignature API.
+- Added support for version v2-21.2.02.00 of the DocuSign ESignature API.
 - Updated the SDK release version.
 
 ## [v2.10.0] - eSignature API v2-21.1.02.00 - 2021-06-04